+++ lib/Db/BookmarkMapper.php
@@ -76,6 +76,9 @@ class BookmarkMapper extends QBMapper {
 	 */
 	private $shareMapper;
 
+	/** @var array<string,int> */
+	private array $userBookmarkCount = [];
+
 	/**
 	 * BookmarkMapper constructor.
 	 *
@@ -154,6 +157,9 @@ class BookmarkMapper extends QBMapper {
 		while ($bookmark = $result->fetchOne()) {
 			$bm = $this->find($bookmark);
 			$this->delete($bm);
+		}
+		if (isset($this->userBookmarkCount[$userId])) {
+			$this->userBookmarkCount[$userId] = 0;
 		}
 		$result->closeCursor();
 	}
@@ -838,6 +844,9 @@ class BookmarkMapper extends QBMapper {
 		);
 
 		$returnedEntity = parent::delete($entity);
+		if (isset($this->userBookmarkCount[$entity->getUserId()])) {
+			$this->userBookmarkCount[$entity->getUserId()] -= 1;
+		}
 
 		$id = $entity->getId();
 
@@ -859,6 +868,7 @@ class BookmarkMapper extends QBMapper {
 		if ($entity->isWebLink()) {
 			$entity->setUrl($this->urlNormalizer->normalize($entity->getUrl()));
 		}
+		$entity->setUrlHash(hash('xxh128', $entity->getUrl()));
 		$entity->setLastmodified(time());
 		parent::update($entity);
 		$this->eventDispatcher->dispatchTyped(new ManipulateEvent('bookmark', $entity->getId()));
@@ -883,19 +893,28 @@ class BookmarkMapper extends QBMapper {
 		if ($entity->isWebLink()) {
 			$entity->setUrl($this->urlNormalizer->normalize($entity->getUrl()));
 		}
-		$entity->setUrlHash(hash('xxh128', $entity->getUrl()));
-
-		if ($entity->getAdded() === null) {
-			$entity->setAdded(time());
+		$entityToInsert = new Bookmark();
+		foreach ($entity->getUpdatedFields() as $field => $_value) {
+			$value = $entity->{'get' . ucfirst($field)}();
+			$entityToInsert->{'set' . ucfirst($field)}($value);
 		}
-		$entity->setLastmodified(time());
-		$entity->setLastPreview(0);
-		$entity->setClickcount(0);
+
+		$entityToInsert->setUrlHash(hash('xxh128', $entity->getUrl()));
+
+		if ($entityToInsert->getAdded() === null) {
+			$entityToInsert->setAdded(time());
+		}
+		$entityToInsert->setLastmodified(time());
+		$entityToInsert->setLastPreview(0);
+		$entityToInsert->setClickcount(0);
 
 		try {
-			parent::insert($entity);
-			$this->eventDispatcher->dispatchTyped(new InsertEvent('bookmark', $entity->getId()));
-			return $entity;
+			parent::insert($entityToInsert);
+			if (isset($this->userBookmarkCount[$entityToInsert->getUserId()])) {
+				$this->userBookmarkCount[$entityToInsert->getUserId()] += 1;
+			}
+			$this->eventDispatcher->dispatchTyped(new InsertEvent('bookmark', $entityToInsert->getId()));
+			return $entityToInsert;
 		} catch (Exception $e) {
 			if ($e->getReason() === Exception::REASON_UNIQUE_CONSTRAINT_VIOLATION) {
 				if ($this->db->inTransaction()) {
@@ -919,12 +938,15 @@ class BookmarkMapper extends QBMapper {
 	public function insertOrUpdate(Entity $entity): Bookmark {
 		try {
 			$newEntity = $this->insert($entity);
-		} catch (AlreadyExistsError $e) {
+		} catch (AlreadyExistsError) {
 			try {
 				$bookmark = $this->findByUrl($entity->getUserId(), $entity->getUrl());
-				$entity->setId($bookmark->getId());
-				$newEntity = $this->update($entity);
-			} catch (DoesNotExistException $e) {
+				foreach ($entity->getUpdatedFields() as $field => $_value) {
+					$value = $entity->{'get' . ucfirst($field)}();
+					$bookmark->{'set' . ucfirst($field)}($value);
+				}
+				$newEntity = $this->update($bookmark);
+			} catch (DoesNotExistException) {
 				$newEntity = $this->insert($entity);
 			}
 		}
@@ -936,6 +958,9 @@ class BookmarkMapper extends QBMapper {
 	 * @throws Exception
 	 */
 	public function countBookmarksOfUser(string $userId) : int {
+		if (isset($this->userBookmarkCount[$userId])) {
+			return $this->userBookmarkCount[$userId];
+		}
 		$qb = $this->db->getQueryBuilder();
 		$qb
 			->select($qb->func()->count('id'))
@@ -947,6 +972,7 @@ class BookmarkMapper extends QBMapper {
 			$count = 0;
 		}
 		$result->closeCursor();
+		$this->userBookmarkCount[$userId] = (int)$count;
 		return (int)$count;
 	}