+++ lib/Migration/Version016002000Date20260201124723.php
@@ -48,35 +48,5 @@ class Version016002000Date20260201124723 extends SimpleMigrationStep {
 	}
 
 	public function postSchemaChange(IOutput $output, Closure $schemaClosure, array $options) {
-		$countQb = $this->db->getQueryBuilder();
-		$countQb->select($countQb->func()->count('id'))->from('bookmarks')->where($countQb->expr()->isNull('url_hash'));
-		$result = $countQb->executeQuery();
-		$count = $result->fetchOne();
-		$output->info('Hashing URLs of n=' . $count . ' bookmarks');
-		$output->startProgress($count);
-
-		$qb = $this->db->getQueryBuilder();
-		$qb->select('id', 'url')->from('bookmarks')->where($qb->expr()->isNull('url_hash'));
-		$result = $qb->executeQuery();
-		$setQb = $this->db->getQueryBuilder();
-		$setQb->update('bookmarks')
-			->set('url_hash', $qb->createParameter('url_hash'))
-			->where($qb->expr()->eq('id', $qb->createParameter('id')));
-		$i = 1;
-		$this->db->beginTransaction();
-		while ($row = $result->fetch()) {
-			if ($i++ % 1000 === 0) {
-				$this->db->commit();
-				$this->db->beginTransaction();
-			}
-			$setQb->setParameter('url_hash', hash('xxh128', $row['url']));
-			$setQb->setParameter('id', $row['id']);
-			$setQb->executeStatement();
-			$output->advance();
-		}
-		$this->db->commit();
-		$output->finishProgress();
-		$result->closeCursor();
 	}
-
 }