+++ CHANGELOG.md
@@ -4,6 +4,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [16.1.3] - 2026-02-21
+
+### Fixed
+* fix(Dashboard): Fix dashboard widgets not loading
+
+
+## [16.1.2] - 2026-02-19
+
+### Fixed
+* fix(Migration): Delete tree entry only if it's a bookmark
+
+## [16.1.1] - 2026-02-19
+
+### Fixed
+* fix(BookmarkMapper): Always update URL hash
+* fix(l10n): Update translations from Transifex
+* fix(BookmarkMapper): Fix bookmark insert logic
+* fix(Migration): Merge per-user-database-duplicate bookmarks
+
 ## [16.1.0] - 2026-02-17
 
 ### New